tutorial/estamp/pyproject.toml
[tool.poetry]
name = "estamp"
version = "0.1.0"
description = ""
authors = ["chengzhenbo <50646861+chengzhenbo@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
pydantic-settings = "^2.0.2"
pillow = "^10.0.0"
pymupdf = "^1.22.5"
img2pdf = "^0.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
